requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord==1.0.1
PyPDF2==1.26.0
python-dotenv==0.15.0
certifi==2019.11.28
pdfjinja==1.1.0
asyncpg==0.21.0
emoji==0.6.0
git+https://github.com/revolunet/pypdftk
